api: remove VideoTrackManager and include it in a generic TrackManager
